moved the test into signals tests, it's not related to continue_as_new
