+++ tests/integ_tests/workflow_tests/continue_as_new.rs
@@ -67,58 +67,3 @@ async fn continue_as_new_multiple_concurrent() {
     }
     worker.run_until_done().await.unwrap();
 }
-
-const SIGNAME: &str = "signame";
-
-async fn continue_as_new_wf_with_sigchan(ctx: WfContext) -> WorkflowResult<()> {
-    let continued_from_execution_run_id =
-        &ctx.workflow_initial_info().continued_from_execution_run_id;
-
-    if continued_from_execution_run_id != "" {
-        Ok(WfExitValue::Normal(()))
-    } else {
-        let mut sigchan = ctx.make_signal_channel(SIGNAME);
-
-        // If we drain the channel with drain_all, we get no signals and the workflow stays stuck.
-        let signals = sigchan.drain_all();
-        info!("Signals received: {:?}", signals);
-
-        // If we drain it with .next(), we get a signal and the workflow gets unstuck.
-        // let signal = sigchan.next().await.unwrap();
-        // info!("Signal received: {:?}", signal);
-
-        return Ok(WfExitValue::continue_as_new(
-            ContinueAsNewWorkflowExecution {
-                arguments: vec![[2].into()].into(),
-                ..Default::default()
-            },
-        ));
-    }
-}
-
-#[tokio::test]
-async fn continue_as_new_with_sigchan() {
-    let wf_name = "continue_as_new_with_sigchan";
-    let mut starter = CoreWfStarter::new(wf_name);
-    starter.worker_config.no_remote_activities(true);
-    let mut worker = starter.worker().await;
-    worker.register_wf(wf_name.to_string(), continue_as_new_wf_with_sigchan);
-
-    let client = starter.get_client().await;
-    let options = SignalWithStartOptions::builder()
-        .task_queue(worker.inner_mut().task_queue())
-        .workflow_id(wf_name)
-        .workflow_type(wf_name)
-        .input(vec![[1].into()].into_payloads().unwrap())
-        .signal_name(SIGNAME)
-        .signal_input(vec![b"tada".into()].into_payloads())
-        .build()
-        .unwrap();
-    let res = client
-        .signal_with_start_workflow_execution(options, WorkflowOptions::default())
-        .await
-        .expect("request succeeds.qed");
-
-    worker.expect_workflow_completion(wf_name, Some(res.run_id));
-    worker.run_until_done().await.unwrap();
-}
